admin-panel/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.6.0
psycopg==3.1.18
psycopg-binary==3.1.18
python-dotenv==1.0.1
split-settings==1.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
uWSGI==2.0.22
requests==2.32.3
pyjwt==2.10.1